refactor: 리팩토링을 위해 필요 없는 파일 삭제
